requirements.txt
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==3.2.8
django-cors-headers==3.8.0
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
exceptiongroup==1.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
outcome==1.2.0
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
pytz==2022.7
redis==4.4.2
selenium==4.7.2
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.14
wsproto==1.2.0
zope.interface==5.5.2
